+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## 0.0.2 - 31-12-2020
+### Added
+- Configuration file (config.vdf)
+- Restart Script without closing the sandbox
+
+### Changed
+- Main class is now 'vulc.luasandbox.Sandbox'
+
 ## 0.0.1 - 29-12-2020
 - Initial release